1. register 창에서 user register, user login 성공했는데, LocalStorage에서 auth_user_token 지우는 현상 수정

+++ www/common/services.remote.js
@@ -91,10 +91,18 @@ angular.module('placekoob.services')
     return deferred.promise;
   }
 
-  function logoutUser() {
-    StorageService.remove('auth_user_token');
-    StorageService.remove('auth_vd_token');
-    StorageService.remove('email');
+  function logoutUser(step) {
+    console.log('Login Step: ' + step);
+    step = step || 0;
+    if (step <= 2){
+      StorageService.remove('auth_user_token');
+    }
+
+    if (step <= 4) {
+      StorageService.remove('email');
+      StorageService.remove('auth_vd_token');
+    }
+
     AppStatus.setUserLogined(false);
   }
